pyproject.toml
[tool.poetry]
name = "agentic-ai-assistant"
version = "0.1.0"
description = "AI assistant using LangChain, RAG, and Gemini"
authors = ["User <user@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "*"
langchain-groq = "*"
python-dotenv = "*"
chromadb = "*"
langchain-community = "*"
langchain-chroma = "*"
langchain-huggingface = "*"
sentence-transformers = "*"
pypdf = "^6.6.0"
langchain-text-splitters = "*"
django = "^5.1"
djangorestframework = "^3.15"
django-cors-headers = "^4.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
